# packages/contracts/ts/openapi.json
{
  "components": {
    "schemas": {
...
      "DevTokenRequest": {
        "description": "Which demo identity to mint a token for.",
        "properties": {
          "sub": {
            "description": "Demo identity 'sub' from data/users.yaml.",
            "minLength": 1,
            "title": "Sub",
            "type": "string"
          }
        },
        "required": [
          "sub"
        ],
        "title": "DevTokenRequest",
        "type": "object"
      },
      "DevTokenResponse": {
        "description": "A freshly minted demo token and the identity it carries.",
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "sub": {
            "title": "Sub",
            "type": "string"
          },
          "tenant": {
            "title": "Tenant",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "sub",
          "tenant",
          "role",
          "purpose",
          "expires_in"
        ],
        "title": "DevTokenResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Signed identity token (ADR-015 stub).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Policy-aware context firewall for production RAG.",
    "title": "ContextGuard",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/dev/token": {
      "post": {
        "description": "Mint a signed token for a demo identity (dev only).",
        "operationId": "mint_token_v1_dev_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mint Token",
        "tags": [
          "dev"
        ]
      }
    },
...
  }
}